Basic Framework of Player Shooting
I added a script to make an empty gameobject follow the mouse's
position. I also added a script that used that gameobject's position to
orient an instantiated weapon gameobject. The final script I added is an
example of how a basic bullet style weapon could work. It takes it's
forward facing position and simply moves along in that direction.