Added the TransitionManager and LoadingSceneController
Added the TransitionManager and LoadingSceneController which handle simple asynchronous transitions between scenes using a simple loading screen
Began refactoring the codebase to use the TransitionManager to handle transitions and to move the scripts into the RedTeam.PewPew namespace
These changes break the game, for now, primarily because the bloated nature of some of the scenes and the large-scale refactoring cause the unity editor to crash while attempting to fix it. Later updates will add further systems like TransitionManager to control the game flow and more efficient solutions for gameplay features and level implementation.